split out NLP tests
